2020, day 1: initial naive solution
This uses the combinations, which is lazy.

A faster way would be to find the difference of each for the pairs and
check if which is present in the list.

A similar method could be used for the triplets, but exclude the sums of
any pairs with a value more than `2020 - min(numbers)`, then check if
that number is in the set of numbers.

There's probably a smarter way still, but I didn't sleep well last
night.

I don't intend on writing everything in Python. I'll consider either
Rust or OCaml for the rest.