2024, day 2 (OCaml)
The awkward bit here was the dampener expression, as I want to avoid
dropping down to using imperative loops where possible so I'm exercising
the right mental muscles.

It bugs me a bit that all the various lists with a dropped entry are
generated, as they Python code is lazy and stops when it finds a
suitable list, but it's not really worth using the Lazy module on such
little data.