2024, day 14, part 2 (prep)
Just some refactoring before I handle part 2. My assumption is that the
robots normally look to be in a high entropy state (disordered), but the
christmas tree expected for part 2 would signal a sudden jump to a low
entropy (ordered) state.

I'm going to experimentally figure out when this is the case. I don't
know how quickly this'll happen (thought I expect it'll take a while),
what threshold to use, but my guess is that I should be able to get a
fair idea from running it over some initial iterations, getting some
information from that on typical Shannon Entropy values, and flag
anything that's outside of a standard deviation of that.

I know that there better ways of doing this, but my statistics knowledge
is pretty rusty. Reading what others have done here will be interesting!