2024, day 12, part 2: it's not elegant, but it works...
There's almost certainly a more elegant way of discovering if a plot has
corners, but this'll do for now.

The principal here is that for a given shape, a side should correspond
to a corner, thus counting corners gives you the number of sides.

A way to improve this would be to take a diagonal and derive the
adjacent edges from this to determine if it's a concave/convex corner or
not. I'll do that later when I feel like cleaning the code up.