2024, day 21, part 2: WIP
There's still a bug in part 2 that crops up when I hit iteration 4. The
numbers I'm getting are far too big, which would tend to indicate that
either the numpad convolution method is incorrect in some way or the
d-pad convolution method is incorrect. I'm certain
`recursive_convolute_dpad`, which would give me the answer for part 2 is
correct, but I need to revisit the earlier parts.

Maybe I need to suck it up and try alternate solutions when generating
the numpad and dpad. It's just that this feels as if it's something that
shouldn't require a search.